% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tgrowth.R
\name{calc_tgrowth}
\alias{calc_tgrowth}
\title{Calculate growth temperature}
\usage{
calc_tgrowth(tmin, tmax, lat, doy)
}
\arguments{
\item{tmin}{Minimum daily temperature (degrees Celsius)}

\item{tmax}{Maximum daily tempearture (degrees Celsius)}

\item{lat}{Latitude (degrees)}

\item{doy}{Day of year (integer between 1 and 365)}
}
\value{
Growth temperature (degrees Celsius), a numeric value. Growth 
temperature is calculated as

\deqn{
T_g = T_\text{max} \left( 1/2 + (1 – x^2)^{1/2}/(2 \cos^{–1} x) \right) 
    + T_\text{min} \left( 1/2 – (1 – x^2)^{1/2}/(2 \cos^{–1} x) \right)
}
with
\deqn{
x = – \tan \lambda \; \tan \delta 
}
where \eqn{\lambda} is latitude and \eqn{\delta} is the solar declination
angle. The solar declination angle is calculated as described in Davis et
al., 2017.
}
\description{
The growth temperature is estimated by approximating the 
diurnal temperature cycle  with a sine curve, where daylight hours are 
determined by the day-of-year and latitude following (Jones, 2013).
}
\examples{
\dontrun{
  tgrowth <- calc_tgrowth(
   tmin = 10.0,
   tmax = 30.0,
   lat = 23.5,
   doy = 180 )
}
}
\references{
Davis, T. W. et al. Simple process-led algorithms for simulating habitats 
(SPLASH v.1.0): robust indices of radiation, evapotranspiration and 
plant-available moisture. Geosci. Model. Dev. 10, 689–708 (2017).

Jones, H. G. (2013). Plants and microclimate: a quantitative approach to 
environmental plant physiology. In Cambridge University Press.
}
