% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vp.R
\name{calc_vp}
\alias{calc_vp}
\title{Calculate vapor pressure from relative humidity}
\usage{
calc_vp(qair = NA, tc = NA, tmin = NA, tmax = NA, patm = NA, elv = NA)
}
\arguments{
\item{qair}{Air specific humidity (g g-1)}

\item{tc}{temperature, deg C}

\item{tmin}{(optional) min daily air temp, deg C}

\item{tmax}{(optional) max daily air temp, deg C}

\item{patm}{Atmospehric pressure (Pa)}

\item{elv}{Elevation above sea level (m) (Used only if \code{patm} is missing 
for calculating it based on standard sea level pressure)}
}
\value{
vapor pressure (Pa)
}
\description{
xxx
}
