% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vp.R
\name{calc_vp_inst}
\alias{calc_vp_inst}
\title{Instantaneous vapour pressure}
\usage{
calc_vp_inst(qair, patm, elv)
}
\arguments{
\item{qair}{Air specific humidity (g g-1)}

\item{patm}{Atmospheric pressure (Pa)}

\item{elv}{Elevation above sea level (m) (Used only if \code{patm} is missing 
for calculating it based on standard sea level pressure)}
}
\value{

}
\description{
Follows Abtew and Meleese (2013), Ch. 5 Vapor Pressure Calculation Methods
}
