% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vpd.R
\name{calc_vpd}
\alias{calc_vpd}
\title{Calculate vapour pressure deficit from relative humidity}
\usage{
calc_vpd(
  qair = NA,
  eact = NA,
  tc = NA,
  tmin = NA,
  tmax = NA,
  patm = NA,
  elv = NA
)
}
\arguments{
\item{qair}{Air specific humidity (g g-1)}

\item{eact}{Water vapour pressure (Pa)}

\item{tc}{temperature, deg C}

\item{tmin}{(optional) min daily air temp, deg C}

\item{tmax}{(optional) max daily air temp, deg C}

\item{patm}{Atmospehric pressure (Pa)}

\item{elv}{Elevation above sea level (m) (Used only if \code{patm} is missing 
for calculating it based on standard sea level pressure)}
}
\value{
vapor pressure deficit (Pa)
}
\description{
Follows Abtew and Meleese (2013), Ch. 5 Vapor Pressure Calculation Methods,
in Evaporation and Evapotranspiration
}
