% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vpd.R
\name{calc_vpd_inst}
\alias{calc_vpd_inst}
\title{Calculate instantenous VPD from ambient conditions}
\usage{
calc_vpd_inst(qair = NA, eact = NA, tc = NA, patm = NA, elv = NA)
}
\arguments{
\item{qair}{Air specific humidity (g g-1)}

\item{eact}{Water vapour pressure (Pa)}

\item{tc}{temperature, deg C}

\item{patm}{Atmospehric pressure (Pa)}

\item{elv}{Elevation above sea level (m) (Used only if \code{patm} is missing 
for calculating it based on standard sea level pressure)}
}
\value{
instantenous VPD from ambient conditions
}
\description{
Follows Abtew and Meleese (2013), Ch. 5 Vapor Pressure Calculation Methods,
in Evaporation and Evapotranspiration
}
