% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_drivers_sofun.R
\name{collect_drivers_sofun}
\alias{collect_drivers_sofun}
\title{Collect all drivers}
\usage{
collect_drivers_sofun(site_info, params_siml, meteo, fapar, co2, params_soil)
}
\arguments{
\item{site_info}{A data frame containing site meta info (rows for sites). 
Required columns are: \code{"sitename", "year_start", 
"year_end", "lon", "lat", "elv"}. See \code{\link{prepare_setup_sofun}} for
details.}

\item{params_siml}{A nested data frame with rows for each site containing 
simulation parameters for SOFUN. See \code{rsofun:run_pmodel_f_bysite} or
\code{rsofun:run_biomee_f_bysite}.}

\item{meteo}{A nested data frame with rows for each site and meteorological
forcing data time series nested inside a column named \code{"data"}.}

\item{fapar}{A nested data frame with rows for each site and fAPAR 
forcing data time series nested inside a column named \code{"data"}.}

\item{co2}{A nested data frame with rows for each site and CO2 
forcing data time series nested inside a column named \code{"data"}.}

\item{params_soil}{Soil texture data descriptor, a data frame with columns
\code{"layer", "fsand", "fclay", "forg" } and \code{"fgravel"}.}
}
\value{
A \code{rsofun} input data frame (see \code{rsofun:p_model_drivers}) for a detailed
description of its structure and contents).
}
\description{
Collect all drivers for site-level simulations 
into a nested data frame with one row for each site.
}
