% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_drivers}
\alias{df_drivers}
\title{Demo data}
\format{
A data frame (tibble)
\describe{
  \item{sitename}{Site name}
  \item{params_siml}{simulation parametrs}
  \item{siteinfo}{site info}
  \item{df_soiltexture}{soil texture info}
  \item{forcing}{climate forcing}
}
}
\usage{
df_drivers
}
\description{
Flux demo data
}
\keyword{datasets}
