% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_header.R
\name{export_header}
\alias{export_header}
\title{Export header dataframe to a temporary file}
\usage{
export_header(header.info, input.source)
}
\arguments{
\item{header.info}{A dataframe containing file header information}

\item{input.source}{A character string of the input file}
}
\value{
Returns nothing, saves header dataframe to a temporary file.
}
\description{
Export header dataframe to a temporary file
}
