% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_interpol.R
\name{gapfill_interpol}
\alias{gapfill_interpol}
\title{Gap fills a modis time series}
\usage{
gapfill_interpol(df, sitename, date_start, date_end, settings)
}
\arguments{
\item{df}{data frame with time series data}

\item{sitename}{a site name}

\item{date_start}{a start year}

\item{date_end}{an end year}

\item{settings}{processing settings file}
}
\value{
gap filled time series
}
\description{
Gap fills a modis time series
}
