% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_gee_bysite.R
\name{gapfill_interpol_gee}
\alias{gapfill_interpol_gee}
\title{Gapfill GEE data products}
\usage{
gapfill_interpol_gee(
  df,
  sitename,
  year_start,
  year_end,
  var_name,
  qc_name,
  prod,
  method_interpol,
  keep
)
}
\arguments{
\item{df}{GEE downloaded data frame}

\item{sitename}{site name}

\item{year_start}{start year}

\item{year_end}{end year}

\item{var_name}{variable}

\item{qc_name}{quality control field}

\item{prod}{product name}

\item{method_interpol}{interpolation method}

\item{keep}{keep (intermediate data??)}
}
\value{
gap filled remote sensing time series
}
\description{
Gap filling routine for remote
sensing data downloaded using the
GEE routine
}
