% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_globalfields.R
\name{get_daily_prec}
\alias{get_daily_prec}
\title{Implements a weather generator}
\usage{
get_daily_prec(mval_prec, mval_wet, set_seed = FALSE, leapyear = FALSE)
}
\arguments{
\item{mval_prec}{A vector of twelve numeric values for monthly values of total precipitation.}

\item{mval_wet}{A vector of twelve integer values for the number of wet days in each month.}

\item{set_seed}{A logical specifying whether a random seed is set.}

\item{leapyear}{A logical specifying whether interpolation is done for a leap year (with 366 days).}
}
\value{
A named list of data frames (tibbles) containing input data for each site is returned.
}
\description{
Implements a weather generator to simulate daily precipitation, given the monthly total and the number of days with rain for each month.
}
