% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings_soilgrids}
\alias{get_settings_soilgrids}
\title{Defines settings for settings for SoilGrids ingest}
\usage{
get_settings_soilgrids(varnam, layer = 1, agg = "mean")
}
\arguments{
\item{varnam}{A charachter string specifying the variable of interest.
See \url{https://www.isric.org/explore/soilgrids/faq-soilgrids#What_do_the_filename_codes_mean} for naming conventions.}

\item{layer}{An integer or vector of integers specifying the soil layers.
See \url{https://www.isric.org/explore/soilgrids/faq-soilgrids#What_do_the_filename_codes_mean} for available layers.
Defaults to \code{"0-5cm"}.}

\item{agg}{A character string specifying the aggregation statistic for depth
 layer-specific values. Defaults to
\code{"mean"}. See \url{https://www.isric.org/explore/soilgrids/faq-soilgrids#What_do_the_filename_codes_mean} for
available statistics.}
}
\value{
A named list containing information required for download 
 from SoilGrids
}
\description{
Defines settings for settings for SoilGrids ingest. Handles the
specification of the required layer identifier.
}
\examples{
\dontrun{settings <- get_settings_soilgrids("soc")}

}
