% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings_wise}
\alias{get_settings_wise}
\title{Defines settings for settings for WISE30sec ingest}
\usage{
get_settings_wise(varnam, layer = 1)
}
\arguments{
\item{varnam}{A charachter string specifying the variable of interest.
See \url{https://www.isric.org/documents/document-type/isric-report-201501-world-soil-property-estimates-broad-scale-modelling} for naming conventions.}

\item{layer}{An integer specifying soil layer.
See \url{https://www.isric.org/documents/document-type/isric-report-201501-world-soil-property-estimates-broad-scale-modelling} for available layers.
Defaults to \code{1}.}
}
\value{
A named list containing information required for download from WISE30sec
}
\description{
Defines settings for settings for WISE30sec ingest. Handles the specification of the required layer identifier.
}
\examples{
\dontrun{settings <- get_settings_wise("CNrt")}

}
