% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmax_fluxnet2015.R
\name{get_tmax_fluxnet2015}
\alias{get_tmax_fluxnet2015}
\title{Get daily maximum temperature for all files in a directory.}
\usage{
get_tmax_fluxnet2015(dir)
}
\arguments{
\item{dir}{A character string specifying the directory in which to look
for site-specific half-hourly data files.}
}
\value{
A list of outputs of the function \link{get_tmax_fluxnet2015}.
}
\description{
Wrapper function to derive daily daily maximum temperature from half-hourly data 
for all site-scale data files in a given directory (argument \code{dir}).
}
\examples{
\dontrun{
 df <- get_tmax_fluxnet2015("./")
}

}
