% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmax_fluxnet2015.R
\name{get_tmax_fluxnet2015_byfile}
\alias{get_tmax_fluxnet2015_byfile}
\title{Get daily maximum temperature}
\usage{
get_tmax_fluxnet2015_byfile(filename_hh, write = FALSE)
}
\arguments{
\item{filename_hh}{A character string specifying the file name containing
site-specific half-hourly (or hourly) data.}

\item{write}{A logical specifying whether daiily daytime tmax should be 
written to a file.}
}
\value{
A data frame (tibble) containing daily daytime tmax.
}
\description{
Derive daily maximum temperature from half-hourly 
data.
}
\examples{
\dontrun{
df <- 
get_tmax_fluxnet2015_byfile(
"./FLX_BE-Vie_FLUXNET2015_FULLSET_HH_1996-2014_1-3.csv"
)
}
}
