% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpd_day_fluxnet2015.R
\name{get_vpd_day_fluxnet2015}
\alias{get_vpd_day_fluxnet2015}
\title{Get daytime VPD for all files in a directory.}
\usage{
get_vpd_day_fluxnet2015(dir)
}
\arguments{
\item{dir}{A character string specifying the directory in which to look
for site-specific half-hourly data files.}
}
\value{
A list of outputs of the function \link{get_vpd_day_fluxnet2015_byfile}.
}
\description{
Wrapper function to derive daily daytime VPD from half-hourly data 
for all site-scale data files in a given directory (argument \code{dir}).
filtering hours where the shortwave incoming radiation (SW_IN_F) is
zero and aggregating taking the mean across remaining hours per day.
}
\examples{
\dontrun{
df <- get_vpd_day_fluxnet2015("./")
}

}
