% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vpd_day_fluxnet2015.R
\name{get_vpd_day_fluxnet2015_byfile}
\alias{get_vpd_day_fluxnet2015_byfile}
\title{Get daytime VPD}
\usage{
get_vpd_day_fluxnet2015_byfile(filename_hh, write = FALSE)
}
\arguments{
\item{filename_hh}{A character string specifying the file name containing
site-specific half-hourly data.}

\item{write}{A logical specifying whether daiily daytime VPD should be 
written to a file.}
}
\value{
A data frame (tibble) containing daily daytime VPD.
}
\description{
Derive daily daytime VPD (vapour pressure deficit) from half-hourly 
data filtering hours where the shortwave incoming radiation (SW_IN_F) 
is greater than zero and aggregating taking the mean across remaining 
hours per day.
}
\examples{
\dontrun{
df <- get_vpd_day_fluxnet2015_byfile(
  "./FLX_BE-Vie_FLUXNET2015_FULLSET_HH_1996-2014_1-3.csv"
  )
}


}
