% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{hash_filename}
\alias{hash_filename}
\title{Sanitize filenames for temporary header files}
\usage{
hash_filename(filename)
}
\arguments{
\item{filename}{}
}
\value{
Returns a character string with problematic characters replaced by '_'
}
\description{
Sanitize filenames for temporary header files
}
