% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_bysite.R
\name{ingest_bysite}
\alias{ingest_bysite}
\title{Data ingest for a single site}
\usage{
ingest_bysite(
  sitename,
  source,
  getvars,
  dir,
  settings = list(),
  timescale = "d",
  year_start = NA,
  year_end = NA,
  lon = ifelse(source == "fluxnet", NA),
  lat = ifelse(source == "fluxnet", NA),
  verbose = FALSE
)
}
\arguments{
\item{sitename}{A character string used as site identification. When data is extracted from
global files or remote servers, \code{sitename} is simply used as a label and any string can
be provided. When data is extraced from site-specific files (e.g. \code{source = "fluxnet"}),
then \code{sitename} is used to identify the file from which data is read.}

\item{source}{A character used as identifiyer for the type of data source
(e.g., \code{"fluxnet"}). See vignette for a full description of available options.}

\item{getvars}{A named list of characters specifying the variable names in
the original source dataset and the variable names in the ingested data frame. Use, e.g.,
\code{getvars = list("gpp" = "GPP_NT_VUT_REF")} to read the variable \code{"GPP_NT_VUT_REF"}
from the original file and convert its name to \code{"gpp"} in the ingested data frame.}

\item{dir}{A character specifying the directory where the data is located.}

\item{settings}{A list of additional, source-specific settings used for reading and processing
original files. Defaults to an empty list which triggers the use of default settings (see
e.g., \link{get_settings_fluxnet}) for \code{source = "fluxnet"}.}

\item{timescale}{A character or vector of characters, specifying the time scale of data used from
the respective source (if multiple time scales are available, otherwise is disregarded). Implemented
time scales are \code{c("d", "m", "y")} for daily, monthly, and yearly, respectively. Defaults
to \code{"d"}.}

\item{year_start}{An integer specifying the first year for which data is to be ingested.}

\item{year_end}{An integer specifying the last year for which data is to be ingested (full years
are read, i.e. all days, or hours, or months in each year).}

\item{lon}{A numeric value specifying the longitude for which data is extraced from global files
or remote data servers. If \code{source = "fluxnet"}, this is not required and set ot \code{NA}.}

\item{lat}{A numeric value specifying the longitude for which data is extraced from global files
or remote data servers. If \code{source = "fluxnet"}, this is not required and set ot \code{NA}.}

\item{verbose}{if \code{TRUE}, additional messages are printed. Defaults to \code{FALSE}.}
}
\value{
A data frame (tibble) containing the time series of ingested data.
}
\description{
Ingests data for a single site and one specific data type, specified by argument \code{source}.
}
\examples{
\dontrun{inputdata <- ingest_bysite()}

}
