% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_campbell.R
\name{ingest_campbell}
\alias{ingest_campbell}
\title{Ingest Campbell Scientific Logger Data.}
\usage{
ingest_campbell(input.source, add.units = TRUE,
  add.measurements = TRUE, header.info = TRUE)
}
\arguments{
\item{input.source}{Character indicating the .dat Campbell Scientific File.}

\item{add.units}{Logical indicating if add.units specified in the data file should be
appended to the end of the variable names specificed in the data file,
defaults to TRUE.}

\item{add.measurements}{Logical indicating if add.measurements type (Avg, Smp,
etc) specified in the data file should be appended to the start of the
variable names specificed in the data file, defaults to TRUE.}

\item{header.info}{A logical indicating if header information is written to a
temporary file that can be restored using ingest_header(input.source).}
}
\value{
This function returns a dataframe containing logger data. If
  header.info = TRUE a temporary file is created for the header data. See
  \code{\link{ingest_header}} for more information.
}
\description{
\code{ingest_campbell} ingests data from Campbell Scientific dataloggers that
are stored in the TAO5 file format with a .dat exentension. \strong{All
ingest functions use the source file's base name as an identifying column to track
provenance and relate data and metadata read from files.}
}
\details{
The TIMESTAMP column will be returned as an POSIXct column.
}
\examples{
campbell_file <- system.file("example_data", "campbell_scientific_tao5.dat", package = "ingestr")
cs_data <- ingest_campbell(input.source = campbell_file)
}
