% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_directory.R
\name{ingest_directory}
\alias{ingest_directory}
\title{Ingest a directory of uniform files}
\usage{
ingest_directory(directory = getwd(),
  ingest.function = utils::read.csv, pattern = "*", collapse = TRUE,
  recursive = FALSE, use.parallel = FALSE, check.duplicates = "warn",
  ...)
}
\arguments{
\item{directory}{A character vector with the name of the directory that
contains your data files. Defaults to the working directory.}

\item{ingest.function}{The function to use to read in the files, defaults to
\code{\link[utils]{read.table}} but can take any ingestr or standard import function.}

\item{pattern}{A character vector providing the pattern to match filenames as
in \code{\link[base]{list.files}}. Defaults to all files "*".}

\item{collapse}{A logical argument, when true a single object is returned,
when false an object is returned for each file. Defaults to \code{TRUE}.}

\item{recursive}{A logical argument, when true files are read recursively,
defaults to \code{TRUE}. See \code{\link[base]{list.files}} for more
information..}

\item{use.parallel}{A logical argument indicating whether the package
\code{\link[parallel]{parallel-package}} should be used.}

\item{check.duplicates}{A character argument specifying the action that
should be taken if files with duplicate contents are detected. One of
"warn", "remove", or NULL to disable checking. Defaults to "warn".}

\item{...}{Additional arguments to pass to the input method}
}
\value{
When \code{collapse = T} a single object matching the output class
  of \code{fun} is returned. When \code{collapse = F} a single object is
  returned matching the output class of \code{fun} in the parent environment
  of the function. The names of the input sources are used as object names
}
\description{
This function reads in all files from a directory using the chosen import
function.  Use the 'pattern' argument to specify a set of files, or a
single file type. If collapse = TRUE \code{\link[dplyr]{bind}} is used
to match column names and bind the imported data into a single object.
\strong{All ingest functions use the source file name as an identifying
column to track provenance and relate data and metadata read from files.
Please check that you have unique file names."}
}
\details{
If \code{check.duplicates = "remove"} then only a single set of records will
be retained when files have identical contents. This does not provide rowwise
checking for duplicates. A separate data.frame specifying the removed
input_source, the number of records removed, and the reason for removal.

If using an \code{ingest_*} function and \code{header.info.name} must be set to
NULL, a character vector of length 1 if \code{collapse = TRUE}, or a named character
vector with length equal to the number of files to be ingested if
\code{collapse = FALSE}. If \code{header.info.name} is set to NULL then
header.info.name will be set to "header_base_directory_name" when
\code{collapse = TRUE} or for each file the header information will be stored
as "header_filename". If a named character vector is supplied the names
should match the file names to be read in and the desired names are stored as
the values.
}
