% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_gee_bysite.R
\name{ingest_gee_bysite}
\alias{ingest_gee_bysite}
\title{Defines settings for settings for Google Earth Engine download}
\usage{
ingest_gee_bysite(
  df_siteinfo,
  start_date,
  end_date,
  overwrite_raw,
  overwrite_interpol,
  band_var,
  band_qc,
  prod,
  prod_suffix,
  varnam,
  productnam,
  scale_factor,
  period,
  python_path,
  gee_path,
  data_path,
  method_interpol,
  keep
)
}
\arguments{
\item{df_siteinfo}{xxx}

\item{start_date}{xxx}

\item{end_date}{xxx}

\item{overwrite_raw}{xxx}

\item{overwrite_interpol}{xxx}

\item{band_var}{xxx}

\item{band_qc}{xxx}

\item{prod}{xxx}

\item{prod_suffix}{xxx}

\item{varnam}{xxx}

\item{productnam}{xxx}

\item{scale_factor}{xxx}

\item{period}{xxx}

\item{python_path}{xxx}

\item{gee_path}{xxx}

\item{data_path}{xxx}

\item{method_interpol}{xxx}

\item{keep}{xxx}
}
\value{
A named list containing information required for download from Google
Earth Engine.
}
\description{
Returns a list of two data frames, one with data at original
modis dates (df), and one interpolated to all days (ddf).
}
\examples{
settings_gee <- get_settings_gee( bundle = "modis_fpar" )

}
