% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_header.R
\name{ingest_header}
\alias{ingest_header}
\title{Ingest Header Data}
\usage{
ingest_header(input.source)
}
\arguments{
\item{input.source}{Character indicating the source file, must match the
input.source used to originally ingest the data}
}
\value{
This function returns a dataframe containing the header information
  for the file. Refer to the help page for \code{ingest_*} for more
  information.
}
\description{
\code{ingest_header} reads in header information from datasets that were
created using \code{ingest_*}. The dataset must have been ingested within the
same R session. \strong{All ingest functions use the source file's base name
as an identifying column to track provenance and relate data and metadata
read from files.}
}
\details{
All \code{ingest_*} functions for data files that include header data write
that data to a temporary file. This header data can then be loaded and
manipulated by the user using \code{ingest_header}.
}
\examples{
campbell_file <- system.file("example_data", "campbell_scientific_tao5.dat", package = "ingestr")
cs_data <- ingest_campbell(input.source = campbell_file)
cs_header <- ingest_header(input.source = campbell_file)
}
