% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iButton.R
\name{ingest_iButton}
\alias{ingest_iButton}
\title{Read in iButton data and parameters}
\usage{
ingest_iButton(file, data = TRUE, metadata = FALSE)
}
\arguments{
\item{file}{Character indicating the raw iButton csv file from 1 Wire}

\item{data}{Logical indicating if the data from the iButton file should be
returned}

\item{metadata}{Logical indicating if the metadata from the iButton file
should be returned}
}
\value{
If data and metadata are both TRUE then a list of data frames with
  items 'data' and 'metadata'. Else a single dataframe of data or metadata
}
\description{
Read in iButton data and parameters
}
