% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_modis_bysite.R
\name{ingest_modis_bysite}
\alias{ingest_modis_bysite}
\title{Defines settings for settings for Google Earth Engine download}
\usage{
ingest_modis_bysite(df_siteinfo, settings)
}
\arguments{
\item{df_siteinfo}{xxx}

\item{settings}{xxx}
}
\value{
A named list containing information required for download from Google
Earth Engine.
}
\description{
Returns a list of two data frames, one with data at original
modis dates (df), and one interpolated to all days (ddf).
}
\examples{
settings_gee <- get_settings_gee( bundle = "modis_fpar" )

}
