% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_north_pacific_gyre.R
\name{ingest_NPGO}
\alias{ingest_NPGO}
\title{Ingest North Pacific Gyre Oscillation data}
\usage{
ingest_NPGO(path = "http://www.o3d.org/npgo/npgo.php", header.info = TRUE,
  header.info.name = "header_npgo")
}
\arguments{
\item{header.info}{A logical indicating if header information is written to a
separate data frame.}

\item{header.info.name}{A character indicating the object name for the
metadata data.frame, defaults to "header_npgo".}
}
\value{
A data frame.  If header.info = TRUE a data.frame is created in the
 parent environment of the function.
}
\description{
\code{ingest_} ingests data from http://www.o3d.org/npgo/ for the North
Pacific Gyre Oscillation. \strong{All ingest functions use the source file
name as an identifying column to track provenance and relate data and metadata
read from files.}
}
\examples{

}
