% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_soilgrids.R
\name{ingest_soilgrids}
\alias{ingest_soilgrids}
\title{SoilGrids point ingest}
\usage{
ingest_soilgrids(siteinfo, settings)
}
\arguments{
\item{settings}{A list returned by a call to \link{get_settings_soilgrids}}

\item{sitename}{A character string specifying the site name (ID)}

\item{lon}{Longitude (numeric)}

\item{lat}{Latitude (numeric)}
}
\value{
A data frame with ingested data, nested in column data.
}
\description{
SoilGrids point data ingest
This follows code outlined here: \url{https://git.wur.nl/isric/soilgrids/soilgrids.notebooks/-/blob/master/markdown/xy_info_from_R.md}
}
\examples{
\dontrun{settings <- ingest_soilgrids("id1", lon = 100, lat = 50)}

}
