% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_spectramax2.R
\name{ingest_spectramaxm2}
\alias{ingest_spectramaxm2}
\title{Ingest SpectraMax M2 Plate Reader Data}
\usage{
ingest_spectramaxm2(input.source, header.info = TRUE)
}
\arguments{
\item{input.source}{A string representing the file to ingest}

\item{header.info}{A logical indicating if header information is written to a separate data frame}
}
\value{
A dataframe. If export.header = TRUE a temporary file is created for
  the header data. See \code{\link{ingest_header}} for more information.
}
\description{
\code{ingest_spectramaxm2} ingests data from raw exports of SpectraMax M2 plate
reader data. It will read in text files with any number of plates exported in
the PlateFormat. \strong{All ingest functions use the source file name as an
identifying column to track provenance and relate data and metadata read from
files. Please check that you have unique file names."}
Header information is extracted for each plate and stored in a data frame.
}
\examples{
*data_description*_file <- system.file("example_data", *Example_File*, package = "ingestr")
ingest_*data_description*(input.source = *data_description*_file)
}
