% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_xle.R
\name{ingest_xle}
\alias{ingest_xle}
\title{Ingest Solinst Levellogger Data from *.xle Files}
\usage{
ingest_xle(input.source, header.info = TRUE,
  collapse.timestamp = FALSE)
}
\arguments{
\item{input.source}{A string representing the file to ingest}

\item{header.info}{A logical indicating if header information is written to a separate data frame}

\item{collapse.timestamp}{A logical indicating if a single timestamp should
returned rather than separate date and time columns. Defaults to FALSE,
returning date and time as character columns}
}
\value{
A dataframe. If export.header = TRUE a temporary file is created for
  the header data. See \code{\link{ingest_header}} for more information.
}
\description{
\code{ingest_xle} ingests data from Solinst levellogers that are stored in
*.xle format. This is the standard format for data offloaded from Solinst
levelloggers and barologgers. \strong{All ingest functions use the source
file name as an identifying column to track provenance and relate data and
metadata read from files. Please check that you have unique file names."}
}
\examples{
xle_file <- system.file("example_data", "solinst.xle", package = "ingestr")
ingest_xle(input.source = xle_file)
}
