% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_dates_dataframe.R
\name{init_dates_dataframe}
\alias{init_dates_dataframe}
\title{Initialises a data frame with dates}
\usage{
init_dates_dataframe(
  yrstart,
  yrend,
  startmoy = 1,
  startdoy = 1,
  timescale = "d",
  endmoy = 12,
  enddom = 31,
  noleap = FALSE
)
}
\arguments{
\item{yrstart}{An integer defining the start year of dates covered by the dataframe}

\item{yrend}{An integer defining the end year of dates covered by the dataframe}

\item{startmoy}{An integer defining the start month of dates covered by the dataframe. Defaults to 1.}

\item{startdoy}{An integer defining the start day-of-year of dates covered by the dataframe. Defaults to 1.}

\item{timescale}{A character string specifying the time steps of dates (in rows). Defaults to \code{"d"}. Any of \code{"y", "m", "d", "h", "hh"}.}

\item{endmoy}{An integer defining the end month of dates covered by the dataframe. Defaults to 12.}

\item{enddom}{An integer defining the end day-of-year of dates covered by the dataframe. Defaults to 31.}

\item{noleap}{Whether leap years are ignored. Defaults to \code{FALSE}.}
}
\value{
A data frame (tibble) with dates.
}
\description{
Creates a dataframe (tibble) with rows for each date (ymd object from
library lubridate) from 'yrstart' to 'yrend'. Intervals of dates is
specified by argument 'freq'.
}
\examples{
ddf <- init_dates_dataframe( 2000, 2003, startmoy=1, startdoy=1, timescale="d", endmoy=12, enddom=31, noleap=FALSE )

}
