% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fapar_ingestr.R
\name{plot_fapar_ingestr}
\alias{plot_fapar_ingestr}
\title{Plot fAPAR for one site}
\usage{
plot_fapar_ingestr(df, settings)
}
\arguments{
\item{df}{The data frame returned by \code{ingest_bysite( source = "gee" )}}

\item{settings}{A list of additional settings used for reading original files.}
}
\value{
A ggplot object.
}
\description{
Visualise the fAPR data filtering, imputing, and interpolation to daily values.
}
\examples{
\dontrun{plot_fapar_ingestr(df, settings)}

}
