% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delimited_imports.R, R/sensor_imports.R
\name{read_iButton}
\alias{read_iButton}
\alias{read_iButton}
\title{Read iButton Read in iButton data and parameters}
\usage{
read_iButton(file, data = TRUE, metadata = FALSE)

read_iButton(file, data = TRUE, metadata = FALSE)
}
\arguments{
\item{file}{Character indicating the raw iButton csv file from 1 Wire}

\item{data}{Logical indicating if the data from the iButton file should be
returned}

\item{metadata}{Logical indicating if the metadata from the iButton file
should be returned}

\item{file}{Character indicating the raw iButton csv file from 1 Wire}

\item{data}{Logical indicating if the data from the iButton file should be
returned}

\item{metadata}{Logical indicating if the metadata from the iButton file
should be returned}
}
\value{
If data and metadata are both TRUE then a list of data frames with
  items 'data' and 'metadata'. Else a single dataframe of data or metadata

If data and metadata are both TRUE then a list of data frames with
  items 'data' and 'metadata'. Else a single dataframe of data or metadata
}
\description{
Read iButton Read in iButton data and parameters

Read iButton Read in iButton data and parameters
}
