% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteinfo_fluxnet2015}
\alias{siteinfo_fluxnet2015}
\title{Meta information of all FLUXNET 2015 sites}
\format{
A data frame (tibble)
\describe{
  \item{mysitename}{Site name}
  \item{lon}{Longitude of site}
}
}
\source{
Based on the file FLX_AA-Flx_BIF_LATEST.csv, distributed via FLUXNET 2015. Additional information for 
Koeppen Geiger classification based on Falge et al. (\url{https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=1530}) and complemented by information given in koeppen-geiger.tif.
}
\usage{
siteinfo_fluxnet2015
}
\description{
User-friendly meta information for all FLUXNET 2015 sites as a data frame with rows for sites and columns for variables.
}
\keyword{datasets}
