% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteinfo_globresp}
\alias{siteinfo_globresp}
\title{Demo data}
\format{
A data frame (tibble)
\describe{
  \item{sitename}{Site name}
  \item{lon}{simulation parametrs}
  \item{lat}{site info}
  \item{elv}{soil texture info}
  \item{year_start}{climate forcing}
  \item{year_end}{climate forcing}
}
}
\usage{
siteinfo_globresp
}
\description{
An example data set to illustrate
All the columns in the `siteinfo_globresp` data.frame are necessary to retrieve
the P-model drivers data: a site name for reference `sitename`, the coordinates
of the site location in degrees `lon` and `lat`, its elevation in meters `elv`,
and the first and last year of observations we want to collect `year_start` and 
`year_end`.
The data describes the location of some sites from Atkin et al. 2017
(the original data is available via the 
[TRY Plant Trait Database](https://www.try-db.org/de/Datasets.php)).
}
\keyword{datasets}
