% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vp.R
\name{calc_vp_inst}
\alias{calc_vp_inst}
\title{Instantaneous vapour pressure}
\usage{
calc_vp_inst(qair, patm)
}
\arguments{
\item{qair}{Air specific humidity (g g-1)}

\item{patm}{Atmospheric pressure (Pa)}
}
\value{
Instantaneous vapour pressure
}
\description{
Follows Abtew and Meleese (2013), Ch. 5 Vapor Pressure Calculation Methods
}
