% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_inputs}
\alias{check_inputs}
\title{Checks Function Inputs for the Correct Class}
\usage{
check_inputs(parameters, class.check)
}
\arguments{
\item{parameters}{A character vector of function parameters that have matching classes}

\item{class.check}{One of "character", "numeric", "logical", "list"
specifying what class \code{parameters} should be}
}
\value{
Returns an error if
}
\description{
Checks Function Inputs for the Correct Class
}
\examples{

}
