% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_inputs}
\alias{check_inputs}
\alias{all_character}
\alias{all_numeric}
\alias{all_logical}
\alias{all_list}
\title{Check function input for the correct class}
\usage{
check_inputs(class.check, ...)

all_character(...)

all_numeric(...)

all_logical(...)

all_list(...)
}
\arguments{
\item{class.check}{One of "character", "numeric", "logical", "list"
specifying what class \code{...} should be.}

\item{...}{Unquoted names of function arguments to check}
}
\value{
Returns an error if the provided parameters do not match the
  specified class.
}
\description{
This function is supplied as a generic form of all_character, all_numeric,
all_logical, and all_list.
}
