% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_modis_bysite.R
\name{gapfill_interpol}
\alias{gapfill_interpol}
\title{Gap fills a modis time series}
\usage{
gapfill_interpol(
  df,
  sitename,
  year_start,
  year_end,
  prod,
  method_interpol,
  keep,
  n_focal
)
}
\arguments{
\item{df}{data frame with time series data}

\item{sitename}{a site name}

\item{year_start}{a start year}

\item{year_end}{an end year}

\item{prod}{a MODIS product}

\item{method_interpol}{the interpolation method}

\item{keep}{keep original data}

\item{n_focal}{average across multiple pixels (if available)}
}
\value{
gap filled time series
}
\description{
Gap fills a modis time series
}
