% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_modis_bysite.R
\name{ingest_modis_bysite}
\alias{ingest_modis_bysite}
\title{Download data by site}
\usage{
ingest_modis_bysite(siteinfo, settings)
}
\arguments{
\item{siteinfo}{site info data frame}

\item{settings}{download settings}
}
\value{
A named list containing information required for download from Google
Earth Engine.
}
\description{
Requires a list of sites, and a settings file.
}
\details{
Returns a list of two data frames, one with data at original
modis dates (df), and one interpolated to all days (ddf).
}
\examples{
\dontrun{
df <- ingest_modis_bysite( bundle = "modis_fpar" )
}
}
