% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_north_pacific_gyre.R
\name{ingest_NPGO}
\alias{ingest_NPGO}
\title{Ingest North Pacific Gyre Oscillation data}
\usage{
ingest_NPGO(input.source = "http://www.o3d.org/npgo/npgo.php",
  export.header = TRUE)
}
\arguments{
\item{input.source}{Character indicating the URI to the HTML representation of the data.}

\item{export.header}{A logical indicating if header information is written to a
separate data frame.}
}
\value{
A data frame.  If export.header = TRUE a temporary file is created for
  the header data. See \code{\link{ingest_header}} for more information.
}
\description{
\code{ingest_} ingests data from http://www.o3d.org/npgo/ for the North
Pacific Gyre Oscillation. \strong{All ingest functions use the source file
name as an identifying column to track provenance and relate data and metadata
read from files.}
}
\examples{
\dontrun{
df_npgo <- ingest_NPGO()  # reads in all the data
}
}
