% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacific_decadal_oscillation.R
\name{ingest_PDO}
\alias{ingest_PDO}
\title{Ingest Pacific Decadal Oscillation data}
\usage{
ingest_PDO(end_year = NULL)
}
\arguments{
\item{end_year}{The last year of data wanted.}
}
\value{
A data frame
}
\description{
Ingest Pacific Decadal Oscillation data
}
