% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delimited_imports.R, R/sensor_imports.R
\name{read_campbell}
\alias{read_campbell}
\alias{read_campbell}
\title{Read Campbell Scientific Data Files}
\usage{
read_campbell(file, units = TRUE, measurement = TRUE)

read_campbell(file, units = TRUE, measurement = TRUE)
}
\arguments{
\item{file}{Character indicating the .dat Campbell Scientific File}

\item{units}{Logical indicating if units should be taken from .dat file}

\item{measurement}{Logical}

\item{file}{Character indicating the .dat Campbell Scientific File}

\item{units}{Logical indicating if units should be taken from .dat file}

\item{measurement}{Logical}
}
\value{
A data frame

A data frame
}
\description{
Read in Campbell Scientific datalogger .dat files

Read in Campbell Scientific datalogger .dat files
}
