% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plumber.R
\name{read_plumber}
\alias{read_plumber}
\title{Reads plumber2 netcdf data}
\usage{
read_plumber(
  site = "AT-Neu",
  path,
  fluxnet_format = FALSE,
  meta_data = FALSE,
  out_path
)
}
\arguments{
\item{site}{fluxnet site name}

\item{path}{path with plumber2 data (both flux and meteo data files)}

\item{fluxnet_format}{convert to fluxnet formatting (TRUE/FALSE)}

\item{meta_data}{return meta-data TRUE/FALSE}

\item{out_path}{where to store the converted data if converted to
fluxnet formatting}
}
\value{
data frame with merged meteo and flux data
}
\description{
Reads data in a given directory
by (fluxnet) site name, optionally
only returns the meta-data of the
site data
}
